+++ ugrid/ugrid.py
@@ -514,7 +514,7 @@ class UGrid:
         if num_faces > 0:
             num_face_nodes_max = np.max(mesh2d.nodes_per_face)
             face_nodes_array = np.full(
-                num_faces * num_face_nodes_max, dtype=np.int, fill_value=-1
+                num_faces * num_face_nodes_max, dtype=np.int32, fill_value=-1
             )
             for index, (face_nodes, nodes_per_face) in enumerate(
                 zip(mesh2d.face_nodes, mesh2d.nodes_per_face)
@@ -629,7 +629,7 @@ class UGrid:
         """
 
         num_edges = len(contacts.mesh1d_indices)
-        edges_array = np.full(num_edges * 2, dtype=np.int, fill_value=-1)
+        edges_array = np.full(num_edges * 2, dtype=np.int32, fill_value=-1)
         for index, (mesh1d_indices, mesh2d_indices) in enumerate(
             zip(contacts.mesh1d_indices, contacts.mesh2d_indices)
         ):
@@ -1098,7 +1098,7 @@ class UGrid:
             byref(c_num_dimensions),
         )
 
-        dimension_vec = np.empty(c_num_dimensions.value, dtype=np.int)
+        dimension_vec = np.empty(c_num_dimensions.value, dtype=np.int32)
         dimension_vec_ptr = as_ctypes(dimension_vec)
         self.__execute_function(
             self.lib.ug_variable_get_data_dimensions,
@@ -1148,7 +1148,7 @@ class UGrid:
 
         data_vec_dimension = functools.reduce(operator.mul, dimension_vec)
 
-        data_vec = np.empty(data_vec_dimension, dtype=np.int)
+        data_vec = np.empty(data_vec_dimension, dtype=np.int32)
         data_vec_ptr = as_ctypes(data_vec)
         variable_name_long = self.__adjust_name(variable_name)
         c_variable_name_encoded = c_char_p(variable_name_long.encode("ASCII"))
